+++ crates/board/CHANGELOG.md
@@ -13,6 +13,8 @@
 
 ### Patch
 
+- Move `Support` bound inside the APIs in `crypto`
+- Fix software crypto support to be transitive
 - Update dependencies
 
 ## 0.4.0
@@ -77,4 +79,4 @@
 
 ## 0.1.0
 
-<!-- Increment to skip CHANGELOG.md test: 10 -->
+<!-- Increment to skip CHANGELOG.md test: 11 -->